OGL: Implement depth access for post-processing shaders
Post shaders that require the depth buffer must declare "RequireDepthBuffer=1" in the configuration block. This is so that an extra resolve is not performed in MSAA mode when not necessary.

Also changes the shader config to initialize config variables in case a shader is missing them, and remove the configuration block from the source that is passed to the GLSL compiler.